All print command in emmc.c are now under preprocessor directive labed DEBUG_LOG except the one in emmcSendData as I was unable to replace it with a delay. Yet to debug: 	getNextClusterFromFat 	readFile

+++ libraries/includes/assemblyFunctions.h
@@ -2,7 +2,7 @@
 #define __ASSEMBLEY_FUNCTIONS_H_
 
 #include <intTypes.h>
-extern void setForeColour(uint32_t);
+extern void setForeColour(uint16_t);
 extern void FullScreenToForeColour();
 extern int frameBufferInit(int, int, int);
 extern void UsbInitialise();
